ai-platform-production-stack/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
numpy==2.1.1
pytest==8.3.2
httpx==0.27.2
faiss-cpu==1.8.0.post1
mlflow==2.16.2
